requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.6
defusedxml==0.6.0
diff-match-patch==20200713
Django==2.2.17
django-cors-headers==3.7.0
django-crispy-forms==1.11.0
django-filter==2.4.0
django-formtools==2.2
django-guardian==2.3.0
django-import-export==2.5.0
django-reversion==3.0.9
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
DjangoUeditor==1.8.143
et-xmlfile==1.0.1
future==0.18.2
httplib2==0.9.2
idna==2.10
importlib-metadata==3.4.0
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.3
Markdown==3.3.3
MarkupPy==1.14
MarkupSafe==1.1.1
mixins==0.1.4
mysqlclient==2.0.3
odfpy==1.4.1
openpyxl==3.0.6
Pillow==8.1.0
pycparser==2.20
pycryptodome==3.10.1
pycryptodomex==3.9.4
PyJWT==1.7.1
PyMySQL==1.0.2
pyOpenSSL==19.1.0
python-alipay-sdk==3.0.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
ronglian-sms-sdk==1.0.0
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.3
xadmin==2.0.1
xlrd==2.0.1
xlwt==1.3.0
zipp==3.4.0
